+++ BioFVM/BioFVM_microenvironment.h
@@ -260,6 +260,7 @@ class Microenvironment
 
 	friend void diffusion_decay_solver__constant_coefficients_LOD_3D( Microenvironment& S, double dt ); 
 	friend void diffusion_decay_solver__constant_coefficients_LOD_2D( Microenvironment& S, double dt ); 
+	friend void diffusion_decay_solver__constant_coefficients_LOD_1D( Microenvironment& S, double dt ); 
 	
 	friend void diffusion_decay_explicit_uniform_rates( Microenvironment& M, double dt );
 	
@@ -281,6 +282,8 @@ extern void diffusion_decay_solver__variable_coefficients_explicit_uniform_mesh(
 
 extern void diffusion_decay_solver__constant_coefficients_LOD_3D( Microenvironment& S, double dt ); 
 extern void diffusion_decay_solver__constant_coefficients_LOD_2D( Microenvironment& S, double dt ); 
+extern void diffusion_decay_solver__constant_coefficients_LOD_1D( Microenvironment& S, double dt ); 
+
 
 extern void diffusion_decay_solver__variable_coefficients_LOD_3D( Microenvironment& S, double dt ); 
 extern void diffusion_decay_solver__variable_coefficients_LOD_2D( Microenvironment& S, double dt );