+++ core/PhysiCell_cell.cpp
@@ -72,8 +72,10 @@
 #include "../BioFVM/BioFVM_vector.h" 
 #include<limits.h>
 
-namespace PhysiCell{
+#include <signal.h>  // for segfault
 
+namespace PhysiCell{
+	
 std::unordered_map<std::string,Cell_Definition*> cell_definitions_by_name; 
 std::unordered_map<int,Cell_Definition*> cell_definitions_by_type; 
 std::vector<Cell_Definition*> cell_definitions_by_index;
@@ -133,7 +135,11 @@ Cell_Definition::Cell_Definition()
 	functions.set_orientation = NULL;
 	
 	cell_definitions_by_index.push_back( this ); 
+	// std::cout << "--- cell_definitions_by_index(1) size= " << cell_definitions_by_index.size() << std::endl;
 	
+	// if (cell_definitions_by_index.size() == 6)
+    // 	raise(SIGSEGV);   // needs #include <signal.h>
+
 	return; 
 }
 
@@ -157,6 +163,7 @@ Cell_Definition::Cell_Definition( Cell_Definition& cd )
 	parameters.pReference_live_phenotype = &phenotype; 
 	
 	cell_definitions_by_index.push_back( this ); 
+	// std::cout << "----- cell_definitions_by_index(2) size= " << cell_definitions_by_index.size() << std::endl;
 	
 	return; 
 }
@@ -180,7 +187,8 @@ Cell_Definition& Cell_Definition::operator=( const Cell_Definition& cd )
 	// this is the whole reason we need ot make a copy constructor 
 	parameters.pReference_live_phenotype = &phenotype; 
 	
-	cell_definitions_by_index.push_back( this ); 
+	// commented out on March 10, 2020 
+	// cell_definitions_by_index.push_back( this ); 
 	
 	return *this; 
 }
@@ -1150,40 +1158,104 @@ void display_cell_definitions( std::ostream& os )
 		Cell_Definition* pCD = cell_definitions_by_index[n]; 
 		os << n << " :: type:" << pCD->type << " name: " << pCD->name << std::endl; 
 
+		// summarize cycle model 
 		if( pCD->phenotype.cycle.pCycle_Model != NULL )
 		{
 			os << "\t cycle model: " << pCD->phenotype.cycle.model().name  
 				<< " (code=" << pCD->phenotype.cycle.model().code << ")" << std::endl; 
-			os << "\t death models: " << std::endl; 
+				
+			// let's show the transition rates 
+			Cycle_Model* pCM = &(pCD->phenotype.cycle.model() ); 
+			Cycle_Data* pCMD = &(pCD->phenotype.cycle.data ); 
+			for( int n=0 ; n < pCM->phases.size() ; n++ )
+			{
+				os << "\t\tPhase " << n << ": " << pCM->phases[n].name << std::endl; 
+			}
+			os << "\t\tCycle transitions: " << std::endl 
+			   << "\t\t-----------------" << std::endl; 
+			for( int n=0 ; n < pCM->phase_links.size() ; n++ )
+			{
+				for( int k=0; k < pCM->phase_links[n].size() ; k++ )
+				{
+					int start = pCM->phase_links[n][k].start_phase_index;
+					int end = pCM->phase_links[n][k].end_phase_index; 
+					os << "\t\t" << pCM->phases[start].name << " --> " 
+						<< pCM->phases[end].name << " w mean duration " 
+						<< 1.0 / pCMD->transition_rate( start,end) << " min" << std::endl; 
+				}
+			}			
+			
 		}
 		else
 		{ 	os << "\t cycle model not initialized" << std::endl; } 
 
+		// summarize death models 
+		os << "\t death models: " << std::endl; 
 		for( int k=0 ; k < pCD->phenotype.death.models.size(); k++ )
 		{
 			os << "\t\t" << k << " : " << pCD->phenotype.death.models[k]->name 
-			<< " (code=" << pCD->phenotype.death.models[k]->code << ")" << std::endl; 
+			<< " (code=" << pCD->phenotype.death.models[k]->code << ")" 
+			<< " with rate " << pCD->phenotype.death.rates[k] << " 1/min" << std::endl; 
 		}
 		
+		// summarize functions 
 		Cell_Functions* pCF = &(pCD->functions); 
 		os << "\t key functions: " << std::endl; 
-		os << "\t\t migration: "; display_ptr_as_bool( pCF->update_migration_bias , std::cout ); 
+		os << "\t\t migration bias rule: "; display_ptr_as_bool( pCF->update_migration_bias , std::cout ); 
 		os << std::endl; 
 		os << "\t\t custom rule: "; display_ptr_as_bool( pCF->custom_cell_rule , std::cout ); 
 		os << std::endl; 
 		os << "\t\t phenotype rule: "; display_ptr_as_bool( pCF->update_phenotype , std::cout ); 
 		os << std::endl; 
+		os << "\t\t volume update function: "; display_ptr_as_bool( pCF->volume_update_function , std::cout ); 
+		os << std::endl; 
+		os << "\t\t mechanics function: "; display_ptr_as_bool( pCF->update_velocity , std::cout ); 
+		os << std::endl; 
+		
+		// summarize motility 
+		
+		Motility* pM = &(pCD->phenotype.motility); 
+		std::string val = "true";
+		if( pM->is_motile == false )
+		{ val = "false"; } 
+	
+		std::string dimen = "3D"; 
+		if( pM->restrict_to_2D == true )
+		{ dimen = "2D"; } 
+
+		os << "\tmotility (enabled: " << val << " in " << dimen << ")" << std::endl 
+			<< "\t\tspeed: " << pM->migration_speed << " micron/min" << std::endl 
+			<< "\t\tbias: " << pM->migration_bias << " " << std::endl 
+			<< "\t\tpersistence time: " << pM->persistence_time << " min" << std::endl 
+			<< "\t\tchemotaxis (enabled: ";
+			
+			val = "true" ;
+			if( pCD->functions.update_migration_bias != chemotaxis_function )
+			{ val = "false"; } 
+		os << val << ")" << std::endl 
+			<< "\t\t\talong " 
+			<< pM->chemotaxis_direction << " * grad(" 
+			<< microenvironment.density_names[ pM->chemotaxis_index ] << ") " << std::endl; 
+			
+		// secretion
+		
+		
+		
+		// mechanics
+		
+		// size 
+	
 		
 		Custom_Cell_Data* pCCD = &(pCD->custom_data); 
 		os << "\tcustom data: " << std::endl; 
 		for( int k=0; k < pCCD->variables.size(); k++)
 		{
-			os << "\t\t" << pCCD->variables[k].name << std::endl; 
+			os << "\t\t" << pCCD->variables[k] << std::endl; 
 		}
 		os << "\tcustom vector data: " << std::endl; 
 		for( int k=0; k < pCCD->vector_variables.size(); k++)
 		{
-			os << "\t\t" << pCCD->vector_variables[k].name << std::endl; 
+			os << "\t\t" << pCCD->vector_variables[k] << std::endl; 
 		}
 		os << "\t\t\tNOTE: custom vector data will eventually be merged with custom data" << std::endl; 
 			
@@ -1236,7 +1308,668 @@ Cell_Definition* find_cell_definition( int search_type )
 	return output; 	
 }
 
+Cell_Definition& get_cell_definition( std::string search_string )
+{
+	// if the registry isn't built yet, then do it! 
+	if( cell_definitions_by_name_constructed == false )
+	{
+		build_cell_definitions_maps(); 
+	}
+	
+	if( cell_definitions_by_name.count( search_string ) > 0 )
+	{ 
+		return *(cell_definitions_by_name.find( search_string )->second ); 
+	} 
+	
+	std::cout << "Warning! Cell_Definition for " << search_string << " not found!" << std::endl; 
+	std::cout << "Returning the default cell definition instead ... " << std::endl; 
+	
+	return cell_defaults; 	
+}
 
+Cell_Definition& get_cell_definition( int search_type )
+{
+	// if the registry isn't built yet, then do it! 
+	if( cell_definitions_by_name_constructed == false )
+	{
+		build_cell_definitions_maps(); 
+	}
+	
+	if( cell_definitions_by_type.count( search_type ) > 0 )
+	{ 
+		return *(cell_definitions_by_type.find( search_type )->second); 
+	} 
+	
+	std::cout << "Warning! Cell_Definition for " << search_type << " not found!" << std::endl; 
+	std::cout << "Returning the default cell definition instead ... " << std::endl; 
+	
+	return cell_defaults; 	
+}
+
+
+
+
+Cell_Definition* initialize_cell_definition_from_pugixml( pugi::xml_node cd_node )
+{
+	Cell_Definition* pCD; 
+	
+	// if this is not "default" then create a new one 
+	if( std::strcmp( cd_node.attribute( "name" ).value() , "default" ) != 0 )
+	{ pCD = new Cell_Definition; }
+	else
+	{ pCD = &cell_defaults; }
+	
+	// set the name 
+	pCD->name = cd_node.attribute("name").value();
+	
+	// set the ID 
+	if( cd_node.attribute("ID" ) )
+	{ pCD->type = cd_node.attribute("ID").as_int(); }
+	else
+	{ pCD->type = -1; } 
+
+	// get the parent definition (if any) 
+	Cell_Definition* pParent = NULL;
+	if( cd_node.attribute( "parent_type" ) )
+	{ pParent = find_cell_definition( cd_node.attribute( "parent_type" ).value() ); }
+	// if it's not the default and no parent stated, inherit from default 
+	if( pParent == NULL && pCD != &cell_defaults )
+	{ pParent = &cell_defaults; } 
+
+	// if we found something to inherit from, then do it! 
+	if( pParent != NULL )
+	{
+		std::cout << "\tInheriting from type " << pParent->name << " ... " << std::endl; 
+		*pCD = *pParent; 
+		
+		// but recover the name and ID (type)
+		pCD->name = cd_node.attribute("name").value();
+		pCD->type = cd_node.attribute("ID").as_int(); 
+	} 
+	
+	// sync to microenvironment
+	pCD->pMicroenvironment = NULL;
+	if( BioFVM::get_default_microenvironment() != NULL )
+	{ pCD->pMicroenvironment = BioFVM::get_default_microenvironment(); }
+
+	// figure out if this ought to be 2D
+	if( default_microenvironment_options.simulate_2D )
+	{
+		std::cout << "Note: setting cell definition to 2D based on microenvironment domain settings ... "
+		<< std::endl; 
+		pCD->functions.set_orientation = up_orientation; 
+		pCD->phenotype.geometry.polarity = 1.0; 
+		pCD->phenotype.motility.restrict_to_2D = true; 
+	}
+	
+	// make sure phenotype.secretions are correctly sized 
+	
+	// pCD->phenotype.secretion.sync_to_current_microenvironment();
+	pCD->phenotype.secretion.sync_to_microenvironment( (pCD->pMicroenvironment) ); 
+	pCD->phenotype.molecular.sync_to_microenvironment( (pCD->pMicroenvironment) );
+	
+	
+	// set the reference phenotype 
+	pCD->parameters.pReference_live_phenotype = &(pCD->phenotype); 
+	pugi::xml_node node = cd_node.child( "phenotype" ); 
+
+	// set up the cell cycle 
+	// make sure the standard cycle models are defined 
+	create_standard_cycle_and_death_models();
+
+	node = node.child( "cycle" ); 
+	if( node )
+	{
+		int model = node.attribute("code").as_int() ; 
+		
+		// Set the model, but only if it was specified. 
+		if( strlen( node.attribute("code").value() ) > 0 )
+		{
+			// set the model 
+			switch( model )
+			{
+				case PhysiCell_constants::advanced_Ki67_cycle_model: 
+					pCD->functions.cycle_model = Ki67_advanced; 
+					break; 
+				case PhysiCell_constants::basic_Ki67_cycle_model: 
+					pCD->functions.cycle_model = Ki67_basic; 
+					break; 
+				case PhysiCell_constants::flow_cytometry_cycle_model: 
+					pCD->functions.cycle_model = flow_cytometry_cycle_model;  
+					break; 
+				case PhysiCell_constants::live_apoptotic_cycle_model: // ?
+					pCD->functions.cycle_model = live;  // ?
+					std::cout << "Warning: live_apoptotic_cycle_model not directly supported." << std::endl		
+							  << "         Substituting live cells model. Set death rates=0." << std::endl; 
+					break; 
+				case PhysiCell_constants::total_cells_cycle_model: 
+					pCD->functions.cycle_model = live; 
+					std::cout << "Warning: total_cells_cycle_model not directly supported." << std::endl		
+							  << "         Substituting live cells model. Set death rates=0." << std::endl; 
+					break; 
+				case PhysiCell_constants::live_cells_cycle_model: 
+					pCD->functions.cycle_model = live; 
+					break; 
+				case PhysiCell_constants::flow_cytometry_separated_cycle_model: 
+					pCD->functions.cycle_model = flow_cytometry_separated_cycle_model; 
+					break; 
+				case PhysiCell_constants::cycling_quiescent_model: 
+					pCD->functions.cycle_model = cycling_quiescent; 
+					break; 
+				default:
+					std::cout << "Warning: Unknown cycle model " << std::endl;
+					exit(-1); 
+					break; 
+			}
+			pCD->phenotype.cycle.sync_to_cycle_model( pCD->functions.cycle_model ); 
+		}
+		
+		// now, if we inherited from another cell, AND 
+		// if that parent type has the same cylce model, 
+		// then overwrite with their transition rates 
+		
+		if( pParent != NULL )
+		{
+			if( pCD->phenotype.cycle.model().code == pParent->phenotype.cycle.model().code )
+			{
+				std::cout << "copying data ... " << std::endl; 
+				std::cout<< pParent->name << " to " << pCD->name << std::endl; 
+				pCD->phenotype.cycle.data = pParent->phenotype.cycle.data; 
+			}
+		}
+		
+		// set the rates 
+		node = node.child( "transition_rates" );
+		if( node )
+		{
+			node = node.child( "rate");
+			while( node )
+			{
+				// which rate 
+				int start = node.attribute("start_index").as_int(); 
+				int end = node.attribute("end_index").as_int(); 
+				// fixed duration? 
+				bool fixed = false; 
+				if( node.attribute( "fixed_duration" ) )
+				{ fixed = node.attribute("fixed_duration").as_bool(); }
+				// actual value of transition rate 
+				double value = xml_get_my_double_value( node ); 
+				
+				// set the transition rate 
+				pCD->phenotype.cycle.data.transition_rate(start,end) = value; 
+				// set it to fixed / non-fixed 
+				pCD->phenotype.cycle.model().phase_link(start,end).fixed_duration = fixed; 
+				
+				node = node.next_sibling( "rate" ); 
+			}
+			
+		}
+	}
+	
+	// here's what it ***should*** do: 
+	// parse the model, get its code 
+	// look for that model 
+	// if the model is not yet there, then add it
+	// otherwise, modify properties of that model 
+	
+	// set up the death models 
+	int death_model_index = 0; 
+	node = cd_node.child( "phenotype" );
+	node = node.child( "death" ); 
+	if( node )
+	{
+		node = node.child( "model" );
+		while( node )
+		{
+			int model = node.attribute("code").as_int() ; 
+			std::cout << "death model: " << model << std::endl; 
+			
+			// check: is that death model already there? 
+			
+			Death* pD = &( pCD->phenotype.death ); 
+			int death_index = pD->find_death_model_index( model );
+			bool death_model_already_exists = false; 
+			if( pD->rates.size() > death_index )
+			{
+				if( pD->models[death_index]->code == model )
+				{ death_model_already_exists = true; } 
+			}
+			
+			// add the death model and its death rate 
+	
+			double rate = xml_get_double_value(node,"rate");
+			
+			// get death model parameters 
+			
+			Death_Parameters death_params; 
+			// if there is a parent and we already found this model, 
+			// start with the inherited parameters 
+			if( death_model_already_exists && pParent != NULL )
+			{
+				death_params = pParent->phenotype.death.parameters[death_index]; 
+			}
+			
+			node = node.child( "parameters" );
+			
+			// only read these parameters if they are specified. 
+			
+			pugi::xml_node node_temp = node.child( "unlysed_fluid_change_rate" );
+			if( node_temp )
+			{ death_params.unlysed_fluid_change_rate = xml_get_my_double_value( node_temp ); }
+
+			node_temp = node.child( "lysed_fluid_change_rate" );
+			if( node_temp )
+			{ death_params.lysed_fluid_change_rate = xml_get_my_double_value( node_temp ); }
+		
+			node_temp = node.child( "cytoplasmic_biomass_change_rate" );
+			if( node_temp )
+			{ death_params.cytoplasmic_biomass_change_rate = xml_get_my_double_value( node_temp ); }
+
+			node_temp = node.child( "nuclear_biomass_change_rate" );
+			if( node_temp )
+			{ death_params.nuclear_biomass_change_rate = xml_get_my_double_value( node_temp ); }
+
+			node_temp = node.child( "calcification_rate" );
+			if( node_temp )
+			{ death_params.calcification_rate = xml_get_my_double_value( node_temp ); }
+
+			node_temp = node.child( "relative_rupture_volume" );
+			if( node_temp )
+			{ death_params.relative_rupture_volume = xml_get_my_double_value( node_temp ); }
+
+			node_temp = node.child( "lysed_fluid_change_rate" );
+			if( node_temp )
+			{ death_params.lysed_fluid_change_rate = xml_get_my_double_value( node_temp ); }
+
+//			death_params.time_units = 
+//				get_string_attribute_value( node, "unlysed_fluid_change_rate", "units" ); 
+			
+			node = node.parent(); 
+					
+			// set the model 
+			// if the model already exists, just overwrite the parameters 
+			switch( model )
+			{
+				case PhysiCell_constants::apoptosis_death_model: 
+//					pCD->phenotype.death.add_death_model( rate , &apoptosis , apoptosis_parameters );
+					if( death_model_already_exists == false )
+					{ pCD->phenotype.death.add_death_model( rate , &apoptosis , death_params ); }
+					else
+					{
+						pCD->phenotype.death.parameters[death_index] = death_params; 
+						pCD->phenotype.death.rates[death_index] = rate; 
+					}
+					std::cout << "apoptosis " << " " << std::endl; 
+					break; 
+				case PhysiCell_constants::necrosis_death_model: 
+					// set necrosis parameters 
+//					pCD->phenotype.death.add_death_model( rate , &necrosis , necrosis_parameters );
+					if( death_model_already_exists == false )
+					{ pCD->phenotype.death.add_death_model( rate , &necrosis , death_params ); }
+					else
+					{
+						pCD->phenotype.death.parameters[death_index] = death_params; 
+						pCD->phenotype.death.rates[death_index] = rate; 						
+					}
+					std::cout << "necrosis " << " "  << std::endl; 
+					break; 
+				case PhysiCell_constants::autophagy_death_model: 
+					std::cout << "Warning: autophagy_death_model not yet supported." << std::endl		
+							  << "         Skipping this model." << std::endl; 
+					break; 
+				default:
+					std::cout << "Warning: Unknown death model " << std::endl;
+					exit(-1); 
+					break; 
+			}
+			
+			// now get transition rates within the death model 
+			// set the rates 
+			node = node.child( "transition_rates" );
+			if( node )
+			{
+				pugi::xml_node node1 = node.child( "rate");
+				while( node1 )
+				{
+					// which rate 
+					int start = node1.attribute("start_index").as_int(); 
+					int end = node1.attribute("end_index").as_int(); 
+					// fixed duration? 
+					bool fixed = false; 
+					if( node1.attribute( "fixed_duration" ) )
+					{ fixed = node1.attribute("fixed_duration").as_bool(); }
+					// actual value of transition rate 
+					double value = xml_get_my_double_value( node1 ); 
+					
+					// set the transition rate 
+					pCD->phenotype.death.models[death_model_index]->transition_rate(start,end) = value; 
+					// set it to fixed / non-fixed 
+					pCD->phenotype.death.models[death_model_index]->phase_link(start,end).fixed_duration = fixed; 
+					
+					node1 = node1.next_sibling( "rate" ); 
+				}
+			}	
+			node = node.parent(); 
+			
+			node = node.next_sibling( "model" ); 
+			death_model_index++; 
+		}
+		
+	}
+	
+	// volume 
+	node = cd_node.child( "phenotype" );
+	node = node.child( "volume" ); 
+	if( node )
+	{
+		Volume* pV = &(pCD->phenotype.volume);
+		
+		pugi::xml_node node_vol = node.child( "total" );
+		if( node_vol )
+		{ pV->total = xml_get_my_double_value( node_vol ); }
+
+		node_vol = node.child( "fluid_fraction" );
+		if( node_vol )
+		{ pV->fluid_fraction = xml_get_my_double_value( node_vol ); }
+		
+		node_vol = node.child( "nuclear" );
+		if( node_vol )
+		{ pV->nuclear = xml_get_my_double_value( node_vol ); }
+
+		node_vol = node.child( "fluid_change_rate" );
+		if( node_vol )
+		{ pV->fluid_change_rate = xml_get_my_double_value( node_vol ); }
+
+		node_vol = node.child( "cytoplasmic_biomass_change_rate" );
+		if( node_vol )
+		{ pV->cytoplasmic_biomass_change_rate = xml_get_my_double_value( node_vol ); }
+
+		node_vol = node.child( "nuclear_biomass_change_rate" );
+		if( node_vol )
+		{ pV->nuclear_biomass_change_rate = xml_get_my_double_value( node_vol ); }
+
+		node_vol = node.child( "calcified_fraction" );
+		if( node_vol )
+		{ pV->calcified_fraction = xml_get_my_double_value( node_vol ); }
+
+		node_vol = node.child( "calcification_rate" );
+		if( node_vol )
+		{ pV->calcification_rate = xml_get_my_double_value( node_vol ); }
+	
+		node_vol = node.child( "relative_rupture_volume" );
+		if( node_vol )
+		{ pV->relative_rupture_volume = xml_get_my_double_value( node_vol ); }
+
+		// set all the parameters to be self-consistent 
+		
+		pV->fluid = pV->fluid_fraction * pV->total; 
+		pV->solid = pV->total-pV->fluid; 
+
+		pV->nuclear_fluid = pV->fluid_fraction * pV->nuclear; 
+		pV->nuclear_solid = pV->nuclear - pV->nuclear_fluid;
+
+		pV->cytoplasmic = pV->total - pV->nuclear;
+		pV->cytoplasmic_fluid = pV->fluid_fraction*pV->cytoplasmic; 
+		pV->cytoplasmic_solid = pV->cytoplasmic - pV->cytoplasmic_fluid; 
+		
+
+		pV->target_solid_cytoplasmic = pV->cytoplasmic_solid;
+		pV->target_solid_nuclear = pV->nuclear_solid;
+		pV->target_fluid_fraction = pV->fluid_fraction;
+		
+		pV->cytoplasmic_to_nuclear_ratio = pV->cytoplasmic / ( 1e-16 + pV->nuclear);
+		pV->target_cytoplasmic_to_nuclear_ratio = pV->cytoplasmic_to_nuclear_ratio; 
+		
+		pV->rupture_volume = pV->relative_rupture_volume * pV->total; // in volume units 
+	}
+	
+	// mechanics 
+	node = cd_node.child( "phenotype" );
+	node = node.child( "mechanics" ); 
+	if( node )
+	{
+		Mechanics* pM = &(pCD->phenotype.mechanics);
+		
+		pugi::xml_node node_mech = node.child( "cell_cell_adhesion_strength" );
+		if( node_mech )
+		{ pM->cell_cell_adhesion_strength = xml_get_my_double_value( node_mech ); }	
+
+		node_mech = node.child( "cell_cell_repulsion_strength" );
+		if( node_mech )
+		{ pM->cell_cell_repulsion_strength = xml_get_my_double_value( node_mech ); }	
+
+		node_mech = node.child( "relative_maximum_adhesion_distance" );
+		if( node_mech )
+		{ pM->relative_maximum_adhesion_distance = xml_get_my_double_value( node_mech ); }	
+
+		node_mech = node.child( "options" );
+		if( node_mech )
+		{
+			pugi::xml_node node_mech1 = node_mech.child( "set_relative_equilibrium_distance" ); 
+			if( node_mech1 )
+			{
+				if( node_mech1.attribute("enabled").as_bool() )
+				{
+					double temp = xml_get_my_double_value( node_mech1 ); 
+					pM->set_relative_equilibrium_distance( temp ); 
+				}
+			}
+
+			node_mech1 = node_mech.child( "set_absolute_equilibrium_distance" ); 
+			if( node_mech1 )
+			{
+				if( node_mech1.attribute("enabled").as_bool() )
+				{
+					double temp = xml_get_my_double_value( node_mech1 ); 
+					pM->set_absolute_equilibrium_distance( pCD->phenotype , temp ); 
+				}
+			}
+		}
+	}
+	
+	// motility 
+	node = cd_node.child( "phenotype" );
+	node = node.child( "motility" ); 
+	if( node )
+	{
+		Motility* pMot = &(pCD->phenotype.motility);
+		
+		pugi::xml_node node_mot = node.child( "speed" );
+		if( node_mot )
+		{ pMot->migration_speed = xml_get_my_double_value( node_mot ); }	
+
+		node_mot = node.child( "migration_bias" );
+		if( node_mot )
+		{ pMot->migration_bias = xml_get_my_double_value( node_mot ); }	
+
+		node_mot = node.child( "persistence_time" );
+		if( node_mot )
+		{ pMot->persistence_time = xml_get_my_double_value( node_mot ); }	
+
+		node_mot = node.child( "options" );
+		if( node_mot )
+		{
+			// enable motility? 
+			pugi::xml_node node_mot1 = node_mot.child( "enabled" ); 
+			pMot->is_motile = xml_get_my_bool_value( node_mot1 ); 
+			
+			// restrict to 2D? 
+			node_mot1 = node_mot.child( "use_2D" ); 
+			pMot->restrict_to_2D = xml_get_my_bool_value( node_mot1 ); 
+			
+			if( default_microenvironment_options.simulate_2D )
+			{
+				std::cout << "Note: Overriding to set cell motility to 2D based on " 
+							<< "microenvironment domain settings ... "
+				<< std::endl; 				
+				pMot->restrict_to_2D = true; 
+			}
+			
+			// automated chemotaxis setup 
+			node_mot1 = node_mot.child( "chemotaxis" ); 
+			if( node_mot1 )
+			{
+				// enabled? if so, set the standard chemotaxis function
+				if( xml_get_bool_value( node_mot1, "enabled" ) )
+				{
+					pCD->functions.update_migration_bias = chemotaxis_function;
+				}	
+				
+				// search for the right chemo index 
+				
+				std::string substrate_name = xml_get_string_value( node_mot1 , "substrate" ); 
+				pMot->chemotaxis_index = microenvironment.find_density_index( substrate_name ); 
+				
+				std::string actual_name = microenvironment.density_names[ pMot->chemotaxis_index ]; 
+				
+				// error check 
+				if( std::strcmp( substrate_name.c_str() , actual_name.c_str() ) != 0 )
+				{
+					std::cout << "Error: attempted to set chemotaxis to \"" 
+						<< substrate_name << "\", which was not found in the microenvironment." << std::endl 
+					<< "       Please double-check your substrate name in the config file." << std::endl << std::endl; 
+					exit(-1); 
+				}
+				
+				// set the direction 
+				
+				pMot->chemotaxis_direction = xml_get_int_value( node_mot1 , "direction" ); 
+				
+				std::cout << pMot->chemotaxis_direction << " * grad( " << actual_name << " )" << std::endl; 
+
+			}
+		}
+	}	
+
+	// secretion
+	
+	node = cd_node.child( "phenotype" );
+	node = node.child( "secretion" ); 
+	if( node )
+	{
+		Secretion* pS = &(pCD->phenotype.secretion);
+		
+		// find the first substrate 
+		pugi::xml_node node_sec = node.child( "substrate" );
+		while( node_sec )
+		{
+			// which substrate? 
+			
+			std::string substrate_name = node_sec.attribute( "name").value(); 
+			int index = microenvironment.find_density_index( substrate_name ); 
+			std::string actual_name = microenvironment.density_names[ index ]; 
+			
+			// error check 
+			if( std::strcmp( substrate_name.c_str() , actual_name.c_str() ) != 0 )
+			{
+				std::cout << "Error: attempted to set secretion/uptake/export for \"" 
+					<< substrate_name << "\", which was not found in the microenvironment." << std::endl 
+				<< "       Please double-check your substrate name in the config file." << std::endl << std::endl; 
+				exit(-1); 
+			}			
+	
+			// secretion rate
+			pugi::xml_node node_sec1 = node_sec.child( "secretion_rate" ); 
+			if( node_sec1 )
+			{ pS->secretion_rates[index] = xml_get_my_double_value( node_sec1 ); }
+			
+			// secretion target 
+			node_sec1 = node_sec.child( "secretion_target" ); 
+			if( node_sec1 )
+			{ pS->saturation_densities[index] = xml_get_my_double_value( node_sec1 ); }
+	
+			// uptake rate 
+			node_sec1 = node_sec.child( "uptake_rate" ); 
+			if( node_sec1 )
+			{ pS->uptake_rates[index] = xml_get_my_double_value( node_sec1 ); }
+			
+			// net export rate 
+			node_sec1 = node_sec.child( "net_export_rate" ); 
+			if( node_sec1 )
+			{ pS->net_export_rates[index] = xml_get_my_double_value( node_sec1 ); }
+			
+			node_sec = node_sec.next_sibling( "substrate" ); 
+		}
+	}	
+
+	
+	// set up custom data 
+	node = cd_node.child( "custom_data" );
+	pugi::xml_node node1 = node.first_child(); 
+	while( node1 )
+	{
+		// name of teh custom data 
+		std::string name = xml_get_my_name( node1 ); 
+		
+		// units 
+		std::string units = node1.attribute( "units").value(); 
+		
+		std::vector<double> values; // ( length, 0.0 ); 
+		
+		// get value(s)
+		std::string str_values = xml_get_my_string_value( node1 ); 
+		csv_to_vector( str_values.c_str() , values ); 
+		
+		// add variable if cell defaults  
+		// if the custom data is not yet found, add it 
+		// first, try scalar 
+		if( values.size() == 1 )
+		{
+			// find the variable 
+			int n = pCD->custom_data.find_variable_index( name ); 
+			// if it exists, overwrite 
+			if( n > -1 )
+			{ pCD->custom_data.variables[n].value = values[0]; }
+			// otherwise, add 
+			else
+			{ pCD->custom_data.add_variable( name, units, values[0] ); }
+		}
+		// or vector 
+		else
+		{ 
+			// find the variable 
+			int n = pCD->custom_data.find_vector_variable_index( name ); 
+			// if it exists, overwrite 
+			if( n > -1 )
+			{ pCD->custom_data.vector_variables[n].value = values; }
+			// otherwise, add 
+			else
+			{ pCD->custom_data.add_vector_variable( name, units, values ); }
+		}
+		
+		node1 = node1.next_sibling(); 
+	}
+	
+	return pCD;
+}
+
+void initialize_cell_definitions_from_pugixml( pugi::xml_node root )
+{
+	pugi::xml_node node = root.child( "cell_definitions" ); 
+	
+	node = node.child( "cell_definition" ); 
+	
+	while( node )
+	{
+		std::cout << "Processing " << node.attribute( "name" ).value() << " ... " << std::endl; 
+		
+		initialize_cell_definition_from_pugixml( node );	
+		build_cell_definitions_maps(); 
+		
+		node = node.next_sibling( "cell_definition" ); 
+	}
+	
+//	build_cell_definitions_maps(); 
+//	display_cell_definitions( std::cout ); 
+	
+	return; 
+}	
+
+void initialize_cell_definitions_from_pugixml( void )
+{
+	initialize_cell_definitions_from_pugixml( physicell_config_root );
+	return; 
+}
 
 
 };