+++ core/PhysiCell_cell.h
@@ -75,6 +75,10 @@
 #include "./PhysiCell_cell_container.h"
 #include "./PhysiCell_constants.h"
 
+#include "../modules/PhysiCell_settings.h" 
+
+#include "./PhysiCell_standard_models.h" 
+
 using namespace BioFVM; 
 
 namespace PhysiCell{
@@ -241,6 +245,13 @@ void build_cell_definitions_maps( void ); // done
 
 Cell_Definition* find_cell_definition( std::string search_string ); // done 
 Cell_Definition* find_cell_definition( int search_type );  
+
+Cell_Definition& get_cell_definition( std::string search_string ); // done 
+Cell_Definition& get_cell_definition( int search_type );  
+
+Cell_Definition* initialize_cell_definition_from_pugixml( pugi::xml_node cd_node ); 
+void initialize_cell_definitions_from_pugixml( pugi::xml_node root ); 
+void initialize_cell_definitions_from_pugixml( void );
 
 };